Added Previous Project button to Nav Bar
Goal is to improve ease of navigation to the users previous project once
they have annotated multiple data clips and want to navigate to the full
list of data clips in this project. Quality of life improvement to
reduce the number of clicks needed to return to the full list of cilps
for the most recently worked on project. If user clicks button and has
no previous project, an error message appears.