bug fix - confident in your annotations button
Confident in your annotations? button previously could not be toggled
back to the non-confident state. It can now be toggled back and forth
and moves the clip in and out of the marked for review file group. Also
changed the text on the button to improve clarity as to what each button
press indicates.